fix: entrypoint
